+++ app/src/main/java/net/gsantner/markor/dialog/RenameDialog.java
@@ -41,7 +41,7 @@ public class RenameDialog extends DialogFragment {
         boolean darkTheme = AppSettings.get().isDarkThemeEnabled();
         dialogBuilder = new AlertDialog.Builder(getActivity(), darkTheme ?
                 R.style.Theme_AppCompat_Dialog : R.style.Theme_AppCompat_Light_Dialog);
-        dialogView = inflater.inflate(R.layout.rename_dialog, null);
+        dialogView = inflater.inflate(R.layout.ui__rename__dialog, null);
 
         dialogBuilder.setTitle(getResources().getString(R.string.rename));
         dialogBuilder.setView(dialogView);